+++ web/src/pages/team/TeamDetailsPage.js
@@ -4,12 +4,12 @@ import PropTypes from "prop-types";
 
 import {Page, Grid, Dimmer, Avatar} from "tabler-react";
 
-import { fetchTeamDetails as fetchTeamDetailsAction } from "../actions/seasons";
-import ShadowBox from "../components/ShadowBox";
+import { fetchTeamDetails as fetchTeamDetailsAction } from "../../actions/seasons";
+import ShadowBox from "../../components/ShadowBox";
 import {FormattedMessage} from "react-intl";
 import moment from "moment";
-import TeamsOnOrganizationList from "../components/organization/AllTeamsOnOrganization";
-import SwitchSeasonButton from "../components/season/SwitchSeasonButton";
+import TeamsOnOrganizationList from "../../components/organization/AllTeamsOnOrganization";
+import SwitchSeasonButton from "../../components/season/SwitchSeasonButton";
 
 class TeamDetailsPage extends React.PureComponent {
   componentDidMount() {
